requirements.txt
asgiref==3.5.2
Django==4.0.6
django-admin-persian-fonts==0.2
django-ckeditor==6.5.0
django-js-asset==2.0.0
django-link-shortener==0.5
six==1.16.0
sqlparse==0.4.2
